Refs 10156. FooType and FooTypeSupport moved into their own files.
