+++ test/unittest/dds/subscriber/FooType.hpp
@@ -1,0 +1,67 @@
+class FooType
+{
+public:
+
+    inline uint32_t index() const
+    {
+        return index_;
+    }
+
+    inline uint32_t& index()
+    {
+        return index_;
+    }
+
+    inline void index(
+            uint32_t value)
+    {
+        index_ = value;
+    }
+
+    inline const std::array<char, 256>& message() const
+    {
+        return message_;
+    }
+
+    inline std::array<char, 256>& message()
+    {
+        return message_;
+    }
+
+    inline void message(
+            const std::array<char, 256>& value)
+    {
+        message_ = value;
+    }
+
+    inline void serialize(
+            eprosima::fastcdr::Cdr& scdr) const
+    {
+        scdr << index_;
+        scdr << message_;
+    }
+
+    inline void deserialize(
+            eprosima::fastcdr::Cdr& dcdr)
+    {
+        dcdr >> index_;
+        dcdr >> message_;
+    }
+
+    inline bool isKeyDefined()
+    {
+        return true;
+    }
+
+    inline void serializeKey(
+            eprosima::fastcdr::Cdr& scdr) const
+    {
+        scdr << index_;
+    }
+
+private:
+
+    uint32_t index_ = 0;
+    std::array<char, 256> message_;
+};
+
